Firmware cleanup, lint, additional CI checks (#10)
* consolidate i2c read/write functions

* update 1024 bytes to use 2048 bytes with PAGE_SECTOR, ordered dependencies by usage, cleanup

* refactor fake off mode, remove led timer start/stop from chip state, update tests

* refactoring timers start/stop

* ignore lint in 2 places

* changed board_get_unique_id param to be uid instead of id for lint

* change name of registers param for lint

* removed underscore prefixes from configureChipState

* add headers to lint check

* rename ms to milliseconds

* undo lint change

* remove functions from setting and mode manager, cleanup, add new test to make sure settings is update in chipState, renaming

* ci to make sure all tests are included, and to make sure there are no lint issues

* show lint errors in ci logs

* lint

* ignore lint from missing driver files, drivers are in .gitignore

* code review

+++ BulbChipSTM32C071FBPx/run_tests.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 
+# Auto-update test runners
+python3 manage_tests.py
+if [ $? -ne 0 ]; then
+    echo "Test management script failed. Aborting."
+    exit 1
+fi
+
 # Create output directory
 mkdir -p Tests/build
 rm -f Tests/build/*
@@ -43,6 +50,10 @@ run_test() {
 echo "Compiling and running test_chip_state..."
 gcc $CFLAGS Tests/test_chip_state.c $UNITY_SRC -o Tests/build/test_chip_state
 run_test ./Tests/build/test_chip_state
+
+echo "Compiling and running test_settings_manager..."
+gcc $CFLAGS Tests/test_settings_manager.c $UNITY_SRC -o Tests/build/test_settings_manager
+run_test ./Tests/build/test_settings_manager
 
 echo "Compiling and running test_mode_manager..."
 gcc $CFLAGS Tests/test_mode_manager.c $UNITY_SRC $LWJSON_SRC Core/Src/json/command_parser.c Core/Src/json/mode_parser.c Core/Src/model/cli_model.c -o Tests/build/test_mode_manager
